// web/src/api/models/evaluation.ts
/* tslint:disable */
/* eslint-disable */
/**
 * ATRAG API
 * ATRAG API Documentation
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { EvaluationStatus } from './evaluation-status';
// May contain unused imports in some cases
// @ts-ignore
import type { LLMConfig } from './llmconfig';

/**
 * 
 * @export
 * @interface Evaluation
 */
export interface Evaluation {
    /**
     * 
     * @type {string}
     * @memberof Evaluation
     */
    'id'?: string;
    /**
     * 
     * @type {string}
     * @memberof Evaluation
     */
    'user_id'?: string;
    /**
     * 
     * @type {string}
     * @memberof Evaluation
     */
    'name'?: string;
    /**
     * 
     * @type {string}
     * @memberof Evaluation
     */
    'collection_id'?: string;
    /**
     * 
     * @type {string}
     * @memberof Evaluation
     */
    'question_set_id'?: string;
    /**
     * 
     * @type {LLMConfig}
     * @memberof Evaluation
     */
    'agent_llm_config'?: LLMConfig;
    /**
     * 
     * @type {LLMConfig}
     * @memberof Evaluation
     */
    'judge_llm_config'?: LLMConfig;
    /**
     * 
     * @type {EvaluationStatus}
     * @memberof Evaluation
     */
    'status'?: EvaluationStatus;
    /**
     * 
     * @type {string}
     * @memberof Evaluation
     */
    'error_message'?: string;
    /**
     * 
     * @type {number}
     * @memberof Evaluation
     */
    'total_questions'?: number;
    /**
     * 
     * @type {number}
     * @memberof Evaluation
     */
    'completed_questions'?: number;
    /**
     * 
     * @type {number}
     * @memberof Evaluation
     */
    'average_score'?: number;
    /**
     * 
     * @type {string}
     * @memberof Evaluation
     */
    'gmt_created'?: string;
    /**
     * 
     * @type {string}
     * @memberof Evaluation
     */
    'gmt_updated'?: string;
}